refactor state serialization, cleanup
