+++ manticore/core/workspace.py
@@ -1,9 +1,7 @@
 import os
-import resource
 import sys
 import glob
 import signal
-import pickle
 import logging
 import tempfile
 import io
@@ -11,8 +9,8 @@ import io
 from contextlib import contextmanager
 from multiprocessing.managers import SyncManager
 
+from manticore.core.state import PickleSerializer
 from .smtlib import solver
-from .smtlib.solver import SolverException
 from .state import State
 
 logger = logging.getLogger(__name__)
@@ -26,44 +24,6 @@ def manager():
         _manager = SyncManager()
         _manager.start(lambda: signal.signal(signal.SIGINT, signal.SIG_IGN))
     return _manager
-
-
-class StateSerializer(object):
-    """
-    StateSerializer can serialize and deserialize :class:`~manticore.core.state.State` objects from and to
-    stream-like objects.
-    """
-
-    def __init__(self):
-        pass
-
-    def serialize(self, state, f):
-        raise NotImplementedError
-
-    def deserialize(self, f):
-        raise NotImplementedError
-
-
-class PickleSerializer(StateSerializer):
-    DEFAULT_RECURSION: int = 0x100000  # 1M
-    MAX_RECURSION: int = 0x1000000  # 16.7M
-
-    def __init__(self):
-        super().__init__()
-        sys.setrecursionlimit(PickleSerializer.DEFAULT_RECURSION)
-
-    def serialize(self, state, f):
-        try:
-            f.write(pickle.dumps(state, 2))
-        except RuntimeError:
-            if sys.getrecursionlimit() >= PickleSerializer.MAX_RECURSION:
-                raise Exception(f'PickleSerializer recursion limit surpassed {PickleSerializer.MAX_RECURSION}, aborting')
-            logger.info(f'Recursion maximum {sys.getrecursionlimit()} hit, increasing')
-            sys.setrecursionlimit(sys.getrecursionlimit() + PickleSerializer.DEFAULT_RECURSION)
-            self.serialize(state, f)
-
-    def deserialize(self, f):
-        return pickle.load(f)
 
 
 class Store(object):
